backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.11.0
pydantic-settings>=2.10.0
python-dotenv>=1.1.0
paho-mqtt>=2.1.0
pytest>=8.4.0
httpx>=0.28.0
